Renamed files, fixed csv writer, and adjusted docstring for python 3
Renamed r3x_reader_p3.py to r3x_reader.py.
Renamed r3x_reader.py to r3x_reader_p2.py.
Python 3 is now the default.

Fixed function to write header data to a csv file using \n line
terminator since the default line terminator is \r\n, which adds a blank
line between each written line.

Confirmed Python 3 functionality. Tested IQ .mat file in SignalVu-PC,
read header .csv file in Excel, confirmed footer .txt file in notepad.
